Moving TestContainers part to another class, so in each test we might not raise new container
