+++ src/test/java/org/example/atm/BankCreationTest.java
@@ -29,21 +29,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @Testcontainers
-
-
-public class BankCreationTest {
-
-    @Container
-    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:15-alpine")
-            .withDatabaseName("testdb")
-            .withUsername("test")
-            .withPassword("test");
-
-    static void configureProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", postgres::getJdbcUrl);
-        registry.add("spring.datasource.username", postgres::getUsername);
-        registry.add("spring.datasource.password", postgres::getPassword);
-    }
+public class BankCreationTest extends AbstractPostgresContainerTest{
 
     @Autowired
     WebApplicationContext context;
@@ -79,7 +65,6 @@ public class BankCreationTest {
                     System.out.println(result.getResponse().getContentAsString());
                     System.out.println("=======================");
                 })
-
                 .andExpect(status().isCreated())
                 .andExpect(jsonPath("$.id").isNumber())
                 .andExpect(jsonPath("$.name").value("AraratBank"))
